move all configuration files to config folder
