+++ server/src/config/db.config.ts
@@ -1,5 +1,5 @@
 import Database from "better-sqlite3";
-import { appEnv } from "../utils/config-env";
+import { appEnv } from "./env.config";
 
 const dbPath = appEnv.DATABASE_URL.replace("sqlite3:", "");
 export const db = new Database(dbPath);
