+++ server/src/config/docs.config.ts
