+++ server/src/db/users.db.ts
@@ -1,5 +1,5 @@
+import { db } from "../config/db.config";
 import { WithDBTimestamps } from "../types/utils";
-import { db } from "./config.db";
 
 export type User = {
   id: string;
