+++ server/src/middlewares/auth.middleware.ts
@@ -3,7 +3,6 @@ import { NextFunction, Request, Response } from "express";
 import { UnauthorizedResponse } from "../utils/response";
 import { verifyAuthToken } from "../utils/auth-token";
 import { findUserById } from "../db/users.db";
-import { db } from "../db";
 
 export function authenticate(req: Request, res: Response, next: NextFunction) {
   const header = req.get("Authorization") || "";
@@ -21,7 +20,7 @@ export function authenticate(req: Request, res: Response, next: NextFunction) {
 
   const decoded = _decoded as jwt.JwtPayload;
 
-  const user = findUserById(db, decoded.payload?.id);
+  const user = findUserById(decoded.payload?.id);
 
   if (!user) {
     return UnauthorizedResponse(res, "Unathorized");
