+++ server/src/utils/asserts.ts
@@ -1,1 +1,3 @@
-export function typeAssert<T extends true>(_t?: T) { }
+export function typeAssert<T extends true>(_t?: T) {
+  return _t;
+}
